refactor: create component mixins
