+++ pages/westmint/create.vue
@@ -3,14 +3,16 @@
     <br />
     <b-tabs v-model="activeTab" destroy-on-hide expanded>
       <b-tab-item v-for="x in components" :key="x" :label="x">
-        <component :is="x" />
+        <component :is="x" @switchToNFT="switchToCreateNFT" />
       </b-tab-item>
     </b-tabs>
   </section>
 </template>
 
 <script lang="ts">
-import { Component, Vue } from 'nuxt-property-decorator'
+import { Component, mixins } from 'nuxt-property-decorator'
+import CreateMixin from '~/utils/mixins/createMixin'
+
 const Collection = () => import('@/components/unique/Create/Create.vue')
 const NFT = () => import('@/components/unique/Create/CreateToken.vue')
 
@@ -36,8 +38,5 @@ const components = { Collection, NFT }
     }
   },
 })
-export default class Remark extends Vue {
-  public activeTab = 0
-  public components: string[] = ['Collection', 'NFT']
-}
+export default class Remark extends mixins(CreateMixin) {}
 </script>
